requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2

Flask==3.0.0
gunicorn==21.2.0
pillow==10.2.0
opencv-python-headless==4.7.0.72
PyYAML==6.0.1
requests==2.32.2
